Move module related code into separate file
